Solve the boards ahead of time
Instead of solving the boards when the user clicks the button, solve it
when creating the board. It'll be a little more annoying when the user
hits one of those slow boards, but I think that's better than having it
be a slow solve.

The user is also given the option to just cancel that one and start a
new board if they want.