feat: Support giving ideas
Sometimes, it can be helpful to give the user a flash of inspiration.
This adds a button to randomly-select a word that will fit the puzzle
for their current state, and then plays it.

The provided ideas are then tracked so that when shared, they'll be
annotated with a little lightbulb (as well as elsewhere in the UI).